requirements.txt
flask
influxdb-client
paho-mqtt
python-dotenv